Coordinates data type set

+++ src/main/java/models/Component.java
@@ -3,9 +3,9 @@ package models;
 public interface Component {
     //void display(); // Display the component on the UI
     String getDetails(); // Return component details for relationships and display
-    void setCoordinates(int x, int y); // Set the position of the component
-    int getX(); // Get the X coordinate
-    int getY();
+    void setCoordinates(double x, double y); // Set the position of the component
+    double getX(); // Get the X coordinate
+    double getY();
     public String getName();
 
 }
